1.02
Worked pretty on updates to the code editing functionalities. Can't even remember everything that I did, but it's working pretty well now. Solve a lot of bugs along the way.

Created a global variable called ControlsDict to keep track of the parameter attributes and thus Tabs (should be called Layers) just needs to keep track of the actual parameter values.