1.05d
Layers now preloads the included and user algorithms into the global scope, making all algorithms available to be used. (No more need to press "load" just to make a function available).

I think there's a new bug now that crops up when a layer's algorithm is changed... I'm pretty sure it's because the algorithm parameters currently only update whenever the object=={}. That allowed me to add an empty layer and assign an algorithm to it, but it's not allowing me to switch.