1.08a
put a decent example algorithm ("example01") into the default project, so that when the application opens, the user is confronted with a nice algorithm to explore and study.

This actually involved a lot of new work, surprisingly. Because there are now multiple algorithms being fetched from the server, I had to do some asynchronous calls in order to get them.