1.09b
Implemented checking and handling for "vestigial" parameters.

Updated the default algorithm.

There still exists an issue when a user adds a parameter to an algorithm at any position other than the end, and doesn't save the project, next time the user opens that project, the new parameter will get added to the end position of the layer object, causing the paramValues argument to the drawFunction() to be out of order. I think the best fix for this is to update the drawFunction to accept an object (with keys) as an argument instead of the array. This will eliminate all the headache of having to enforce strict order.