1.09c
Changed the drawFunctions arguments list to include object instead of paramValues. This is much more robust, because with the array paramValues, the order of the values was paramount, and there were many ways in which the order could becoe compromised. Passing the object, on the other hand, utilizes key-value mapping, so it's immune to order.