1.11a
Updated the "draw_" function so that it only requires the argument object and not (object, layer, ctx) as it did before. This entailed placing the ctx information into a global variable called ctxToDrawToNow. This variable is accessible by the primitive drawing functions drawRect() and drawCircle() so that Layers and is updated in each drawTab() request so that Layers always knows where to draw stuff.